Couple of fixes mainly for avro records - fixes after review
